+++ components/event-card.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link'
 import tt from 'tinytime'
 import { past } from '../lib/past'
 
-const Event = ({
+const EventCard = ({
   name,
   slug,
   start,
@@ -43,4 +43,4 @@ const Event = ({
   </Link>
 )
 
-export default Event
+export default EventCard
